# Docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HR招聘系统 API",
    "version": "1.0.0",
    "description": "\n智能招聘管理系统后端API文档\n\n## 功能模块\n- **岗位设置** - 岗位标准管理、简历分配\n- **简历筛选** - 简历上传与AI初筛\n- **视频分析** - 面试视频分析（预留）\n- **面试辅助** - AI面试问答助手\n- **最终推荐** - 候选人综合评估\n    "
  },
  "paths": {
...
    "/api/positions/{position_id}/": {
      "get": {
        "operationId": "positions_retrieve_2",
        "description": "获取指定岗位的详细信息，包含关联简历",
        "summary": "获取岗位详情",
        "parameters": [
          {
            "in": "query",
            "name": "include_resumes",
            "schema": {
              "type": "boolean"
            },
            "description": "是否包含关联简历"
          },
          {
            "in": "path",
            "name": "position_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "tags": [
          "positions"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "basicAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPositionDetailResp"
                }
              }
            },
            "description": ""
          }
        }
      },
      "put": {
        "operationId": "positions_update",
        "description": "更新指定岗位的信息",
        "summary": "更新岗位",
        "parameters": [
          {
            "in": "path",
            "name": "position_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "tags": [
          "positions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreateRequestRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreateRequestRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreateRequestRequest"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "basicAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPositionUpdateResp"
                }
              }
            },
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "positions_destroy",
        "description": "软删除指定岗位",
        "summary": "删除岗位",
        "parameters": [
          {
            "in": "path",
            "name": "position_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "tags": [
          "positions"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "basicAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDeleteResponse"
                }
              }
            },
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiPositionDetailResp": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PositionDetail"
          },
          "code": {
            "type": "integer",
            "default": 200,
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "default": "成功",
            "description": "消息"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "ApiPositionUpdateResp": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PositionItem"
          },
          "code": {
            "type": "integer",
            "default": 200,
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "default": "成功",
            "description": "消息"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "PositionCreateRequestRequest": {
        "type": "object",
        "description": "创建岗位请求",
        "properties": {
          "position": {
            "type": "string",
            "minLength": 1,
            "description": "岗位名称"
          },
          "department": {
            "type": "string",
            "minLength": 1,
            "description": "部门"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "description": "岗位描述"
          },
          "required_skills": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "必需技能"
          },
          "optional_skills": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "可选技能"
          },
          "min_experience": {
            "type": "integer",
            "description": "最低经验年限"
          },
          "education": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "学历要求"
          },
          "certifications": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "证书要求"
          },
          "salary_range": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "薪资范围"
          },
          "project_requirements": {
            "description": "项目要求"
          }
        },
        "required": [
          "position"
        ]
      },
      "PositionDeleteResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "default": 200
          },
          "message": {
            "type": "string",
            "default": "成功"
          },
          "data": {
            "nullable": true
          }
        }
      },
      "PositionDetail": {
        "type": "object",
        "description": "岗位详情（含简历）",
        "properties": {
          "id": {
            "type": "string",
            "description": "岗位ID"
          },
          "position": {
            "type": "string",
            "description": "岗位名称"
          },
          "department": {
            "type": "string",
            "description": "部门"
          },
          "description": {
            "type": "string",
            "description": "岗位描述"
          },
          "required_skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "必需技能"
          },
          "optional_skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "可选技能"
          },
          "min_experience": {
            "type": "integer",
            "description": "最低经验年限"
          },
          "education": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "学历要求"
          },
          "certifications": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "证书要求"
          },
          "salary_range": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "薪资范围"
          },
          "project_requirements": {
            "description": "项目要求"
          },
          "resume_count": {
            "type": "integer",
            "description": "简历数量"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "resumes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PositionResume"
            },
            "description": "关联简历"
          }
        },
        "required": [
          "certifications",
          "created_at",
          "department",
          "description",
          "education",
          "id",
          "min_experience",
          "optional_skills",
          "position",
          "project_requirements",
          "required_skills",
          "resume_count",
          "salary_range"
        ]
      },
      "PositionItem": {
        "type": "object",
        "description": "岗位项",
        "properties": {
          "id": {
            "type": "string",
            "description": "岗位ID"
          },
          "position": {
            "type": "string",
            "description": "岗位名称"
          },
          "department": {
            "type": "string",
            "description": "部门"
          },
          "description": {
            "type": "string",
            "description": "岗位描述"
          },
          "required_skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "必需技能"
          },
          "optional_skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "可选技能"
          },
          "min_experience": {
            "type": "integer",
            "description": "最低经验年限"
          },
          "education": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "学历要求"
          },
          "certifications": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "证书要求"
          },
          "salary_range": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "薪资范围"
          },
          "project_requirements": {
            "description": "项目要求"
          },
          "resume_count": {
            "type": "integer",
            "description": "简历数量"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        },
        "required": [
          "certifications",
          "created_at",
          "department",
          "description",
          "education",
          "id",
          "min_experience",
          "optional_skills",
          "position",
          "project_requirements",
          "required_skills",
          "resume_count",
          "salary_range"
        ]
      },
...
      "PositionResume": {
        "type": "object",
        "description": "岗位关联的简历",
        "properties": {
          "id": {
            "type": "string",
            "description": "简历ID"
          },
          "candidate_name": {
            "type": "string",
            "description": "候选人姓名"
          },
          "position_title": {
            "type": "string",
            "description": "应聘岗位"
          },
          "resume_content": {
            "type": "string",
            "description": "简历内容"
          },
          "screening_score": {
            "nullable": true,
            "description": "筛选得分"
          },
          "screening_summary": {
            "type": "string",
            "nullable": true,
            "description": "筛选摘要"
          },
          "report_md_url": {
            "type": "string",
            "nullable": true,
            "description": "MD报告URL"
          },
          "report_json_url": {
            "type": "string",
            "nullable": true,
            "description": "JSON报告URL"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        },
        "required": [
          "candidate_name",
          "created_at",
          "id",
          "position_title",
          "report_json_url",
          "report_md_url",
          "resume_content",
          "screening_score",
          "screening_summary"
        ]
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    }
  },
...
}